Added worker threads option.
Now with the command line option -w you can specify how many worker threads to use. This is a simple first pass at handling better concurrency. ebb_rails defaults to 1 worker.

+++ benchmark/server_test.rb
@@ -80,7 +80,7 @@ class ServerTest
   end
   
   def kill
-    Process.kill('KILL', @pid)
+    Process.kill('KILL', @pid) if @pid
   end
   
   def running?
